fix validator page and refactor lcd requests
