+++ src/containers/application/notFound.jsx
@@ -1,7 +1,7 @@
 import styles from './notFound.module.scss';
 
 function NotFound({ text }) {
-  return <div className={styles.container} />;
+  return <div className={styles.container}>{text || 'not found'}</div>;
 }
 
 export default NotFound;
