+++ src/containers/energy/component/actionBar.tsx
@@ -5,6 +5,7 @@ import { Link } from 'react-router-dom';
 import { useSigningClient } from 'src/contexts/signerClient';
 import { PATTERN_CYBER, PATTERN_CYBER_CONTRACT } from 'src/constants/patterns';
 import { DEFAULT_GAS_LIMITS } from 'src/constants/config';
+import { getTxs } from 'src/services/transactions/lcd';
 import {
   Dots,
   ActionBarContentText,
@@ -17,7 +18,6 @@ import {
   Input,
 } from '../../../components';
 import { LEDGER } from '../../../utils/config';
-import { getTxs } from '../../../utils/search/utils';
 import { ValueImg } from '../ui';
 import { routes } from '../../../routes';
