+++ src/containers/ipfs/hooks/useGetCreator.ts
@@ -1,7 +1,53 @@
 import { useQuery } from '@tanstack/react-query';
 import { CreatorCyberLink } from 'src/types/cyberLink';
 import { Option } from 'src/types';
-import { getCreator } from '../../../utils/search/utils';
+import { getTransactions } from 'src/services/transactions/lcd';
+
+// TODO: refactor this
+const getCreator = async (cid) => {
+  try {
+    const response = await getTransactions({
+      events: [
+        {
+          key: 'cyberlink.particleTo',
+          value: cid,
+        },
+      ],
+      pagination: {
+        limit: 1,
+        offset: 0,
+      },
+    });
+
+    const response2 = await getTransactions({
+      events: [
+        {
+          key: 'cyberlink.particleFrom',
+          value: cid,
+        },
+      ],
+      pagination: {
+        limit: 1,
+        offset: 0,
+      },
+    });
+
+    const h1 = Number(response.txResponses?.[0]?.height || 0);
+    const h2 = Number(response2.txResponses?.[0]?.height || 0);
+
+    if (h1 === 0) {
+      return response2;
+    }
+    if (h2 === 0) {
+      return response;
+    }
+
+    return h1 < h2 ? response : response2;
+  } catch (error) {
+    console.log(error);
+    return null;
+  }
+};
 
 function useGetCreator(cid: string) {
   const { data } = useQuery(
