+++ src/features/rank/index.md
@@ -1,0 +1,3 @@
+Move `rank` files here
+
+Same to https://github.com/cybercongress/go-cyber/tree/main/x/rank module
