+++ src/pages/Keys/ActionBar/actionBarKeplr.tsx
@@ -12,8 +12,9 @@ import {
 } from 'src/components';
 import { LEDGER } from 'src/utils/config';
 import { PATTERN_CYBER } from 'src/constants/patterns';
-import { getTxs } from 'src/utils/search/utils';
+
 import { DEFAULT_GAS_LIMITS, BASE_DENOM } from 'src/constants/config';
+import { getTxs } from 'src/services/transactions/lcd';
 
 const { STAGE_ERROR, STAGE_SUBMITTED, STAGE_CONFIRMING, STAGE_CONFIRMED } =
   LEDGER;
