+++ src/pages/robot/_refactor/account/hooks/useGetCommunity.PREVIOUS.ts
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react';
 import useQueueIpfsContent from 'src/hooks/useQueueIpfsContent';
 
-import { getFollowers, getFollows } from '../../../../../utils/search/utils';
 import { getIpfsHash } from 'src/utils/ipfs/helpers';
 import { PATTERN_CYBER } from 'src/constants/patterns';
 import { useBackend } from 'src/contexts/backend/backend';
